Finished Section 9: Authorization.

+++ backend/routes/posts.js
@@ -32,6 +32,7 @@ router.post('', checkAuth, multer({storage: storage}).single("image"), (req, res
         title: req.body.title,
         content: req.body.content,
         imagePath: url + '/images/' + req.file.filename,
+        creator: req.userData.userId,
     });
     post.save().then(createdPost => {
         res.status(201).json({
@@ -57,13 +58,20 @@ router.patch('/:id', checkAuth, multer({storage: storage}).single("image"), (req
         title: req.body.title,
         content: req.body.content,
         imagePath: imagePath,
+        creator: req.userData.userId,
     });
 
-    Post.updateOne({_id: req.params.id}, post).then(createdPost => {
-        res.status(200).json({
-            message: "Post updated sucesfully!",
-            postId: createdPost._id,
-        });
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId, }, post).then(result => {
+        if (result.modifiedCount > 0){
+            res.status(200).json({
+                message: "Post updated sucesfully!",
+                postId: result._id,
+            });
+        }else{
+            res.status(401).json({
+                message: "Not authorized!",
+            });
+        }
     });
 })
 
@@ -107,10 +115,16 @@ router.get('/:id', (req, res, next) => {
 });
 
 router.delete('/:id', checkAuth,  (req, res, next) => {
-    Post.deleteOne({_id: req.params.id}).then(result => {
-        res.status(200).json({
-            message: "Post deleted sucesfully!",
-        });
+    Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+        if (result.deletedCount > 0){
+            res.status(200).json({
+                message: "Post deleted sucesfully!",
+            });
+        }else{
+            res.status(401).json({
+                message: "Not authorized!",
+            });
+        }
     });
 });
